Workaround for Baumer GenTL GEV interface enumeration bug.
The bug leads to unbounded allocation of memory upon calling TLUpdateInterfaceList().
As a temporary workaround, interfaces will only be enumerated once and kept open
until all systems resources are cleared by System::clearSystems() call. The drawback
of this workaround is that the list of interfaces will be wrong if interfaces are
created or deleted after first enumeration.